Refactor initializeApp call, components to separate variables
